Moves filter buttons to other component
