+++ src/signals/incident-management/containers/IncidentOverviewPage/index.js
@@ -1,9 +1,10 @@
-import React, { Fragment, useEffect } from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import { compose, bindActionCreators } from 'redux';
-import { Row, Column } from '@datapunt/asc-ui';
+import { Row, Column, Button } from '@datapunt/asc-ui';
+import { disablePageScroll, enablePageScroll } from 'scroll-lock';
 
 import PageHeader from 'containers/PageHeader';
 import injectSaga from 'utils/injectSaga';
@@ -14,13 +15,16 @@ import {
   makeSelectCategories,
 } from 'containers/App/selectors';
 import LoadingIndicator from 'shared/components/LoadingIndicator';
+import Filter from 'signals/incident-management/containers/Filter';
+import Modal from 'components/Modal';
 import makeSelectOverviewPage, { makeSelectIncidentsCount } from './selectors';
 import reducer from './reducer';
 import saga from './saga';
-
 import { requestIncidents, incidentSelected } from './actions';
 import ListComponent from './components/List';
 import Pager from './components/Pager';
+
+let lastActiveElement = null;
 
 export const IncidentOverviewPageContainerComponent = ({
   onRequestIncidents,
@@ -28,6 +32,35 @@ export const IncidentOverviewPageContainerComponent = ({
   incidentsCount,
   onIncidentSelected,
 }) => {
+  const [modalIsOpen, toggleModal] = useState(false);
+
+  const openModal = () => {
+    disablePageScroll();
+    toggleModal(true);
+    lastActiveElement = document.activeElement;
+  };
+
+  function closeModal() {
+    enablePageScroll();
+    toggleModal(false);
+    lastActiveElement.focus();
+  }
+
+  useEffect(() => {
+    const escFunction = (event) => {
+      /* istanbul ignore else */
+      if (event.keyCode === 27) {
+        closeModal();
+      }
+    };
+
+    document.addEventListener('keydown', escFunction);
+
+    return () => {
+      document.removeEventListener('keydown', escFunction);
+    };
+  });
+
   useEffect(() => {
     onRequestIncidents({});
   }, []);
@@ -36,7 +69,20 @@ export const IncidentOverviewPageContainerComponent = ({
 
   return (
     <Fragment>
-      <PageHeader />
+      <PageHeader>
+        <Button
+          data-testid="modalBtn"
+          type="button"
+          color="primary"
+          onClick={openModal}
+        >
+          Filteren
+        </Button>
+
+        <Modal isOpen={modalIsOpen} onClose={closeModal} title="Filters">
+          <Filter onSubmit={closeModal} onCancel={closeModal} />
+        </Modal>
+      </PageHeader>
 
       <Row>
         <Column span={12} wrap>
